+++ src/routes/_appRoot.notes_.$.tsx
@@ -332,6 +332,7 @@ function NotePage() {
                   onToggleCollapse={touch}
                   startEditing={isNewNote && !showsTitle}
                   readOnly={readOnlyShare}
+                  browse={readOnlyShare}
                   highlightHeading={highlightHeading}
                   onExitTop={() => setTitleFocusSignal((n) => n + 1)}
                   focusFirstSignal={focusFirstSignal}